refactor input handling
Games should expose the actions rather than input commands. A function
to convert actions to input commands is provided. Actions, unlike
input commands, are uniform: an action is accompanied by a float32,
regardless of whether it's a binary action or not, making them easy
to handle by various input systems like Steam Input and (in the
future) SDL_ActionSet.